requirements.txt
ecdsa==0.16.1
requests
fastapi
uvicorn
websocket-client